+++ src/module.ts
@@ -1,11 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { DeleteComponent } from './components/delete';
-import { FormComponent } from './components/form';
-import { ListComponent } from './components/list';
-import { ReadComponent } from './components/read';
-
 /**
  * The module containing all of the Cruddle components.
  * This should be imported into your relevant application
@@ -15,17 +10,9 @@ import { ReadComponent } from './components/read';
   imports: [CommonModule],
 
   declarations: [
-    DeleteComponent,
-    FormComponent,
-    ListComponent,
-    ReadComponent
   ],
 
   exports: [
-    DeleteComponent,
-    FormComponent,
-    ListComponent,
-    ReadComponent
   ]
 })
 export class CruddleModule {}